+++ config/config.exs
@@ -137,9 +137,7 @@ config :lti_1p3,
   ags_line_item_prefix: "oli-torus-"
 
 config :ex_aws,
-  access_key_id: [{:system, "AWS_ACCESS_KEY_ID"}, :instance_role],
-  secret_access_key: [{:system, "AWS_SECRET_ACCESS_KEY"}, :instance_role],
-  region: System.get_env("AWS_REGION", "us-east-1")
+  region: {:system, "AWS_REGION"}
 
 # Configures Elixir's Logger
 config :logger, :console,